+++ packages/browser/src/index.ts
@@ -20,11 +20,17 @@ export {
   captureEvent,
   captureMessage,
   configureScope,
-  withScope,
   getHubFromCarrier,
   getCurrentHub,
   Hub,
   Scope,
+  setContext,
+  setExtra,
+  setExtras,
+  setTag,
+  setTags,
+  setUser,
+  withScope,
 } from '@sentry/core';
 
 export { BrowserOptions } from './backend';